extract area-of-study and student into web/modules
