+++ modules/web/modules/area-of-study/area-of-study.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.scss';
+@import '../../styles/variables.scss';
 
 .area {
 	&.errored {
