+++ modules/web/modules/area-of-study/expression.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/variables.scss';
-@import '../../../styles/mixins.scss';
+@import '../../styles/variables.scss';
+@import '../../styles/mixins.scss';
 @import './variables-expression.scss';
 
 .expression {
