+++ modules/web/modules/area-of-study/index.js
@@ -1,0 +1,1 @@
+export { default } from './area-of-study'
