+++ modules/web/modules/area-of-study/variables-expression.scss
