+++ modules/web/modules/student/area-of-study-group.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/variables.scss';
-@import '../../../styles/mixins.scss';
+@import '../../styles/variables.scss';
+@import '../../styles/mixins.scss';
 
 .area-of-study-group {
 	@include card();
