+++ modules/web/modules/student/area-of-study-sidebar.js
@@ -15,10 +15,10 @@ import union from 'lodash/union'
 import uniq from 'lodash/uniq'
 import values from 'lodash/values'
 
-import { sortStudiesByType } from '../../../../object-student/sort-studies-by-type'
+import { sortStudiesByType } from '../../../object-student/sort-studies-by-type'
 import AreaOfStudyGroup from './area-of-study-group'
-import Button from '../../../components/button'
-import { areaTypeConstants } from '../../../../object-student/area-types'
+import Button from '../../components/button'
+import { areaTypeConstants } from '../../../object-student/area-types'
 
 export default function AreaOfStudySidebar(props) {
 	const { allAreas, student, showAreaPickerFor } = props
