+++ modules/web/modules/student/area-picker.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/variables.scss';
-@import '../../../styles/mixins.scss';
+@import '../../styles/variables';
+@import '../../styles/mixins';
 
 .add-area {
 	padding: $area-edge-padding;
