+++ modules/web/modules/student/graduation-status.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.scss';
+@import '../../styles/variables.scss';
 
 .graduation-status {
 	font-size: 0.85em;
