+++ modules/web/modules/student/student.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.scss';
+@import '../../styles/variables.scss';
 
 .student {
 	height: 100vh;
