+++ modules/web/routes/create/routes/sis/sis-import.js
@@ -11,7 +11,7 @@ import {
 } from '../../../../../school-st-olaf-college'
 import { BrowserExtensionsComponent } from '../../../../components/browser-extensions'
 import { getCourse } from '../../../../helpers/get-courses'
-import StudentSummary from '../../../student/components/student-summary'
+import StudentSummary from '../../../../modules/student/student-summary'
 import map from 'lodash/map'
 import groupBy from 'lodash/groupBy'
 import sortBy from 'lodash/sortBy'
