+++ modules/web/routes/student/index.js
@@ -17,7 +17,7 @@ export default {
 
 	getComponents(location, cb) {
 		require.ensure([], () => {
-			cb(null, { content: require('./containers/student').default })
+			cb(null, { content: require('./../../modules/student').default })
 		}, 'student.components')
 	},
 }
